appium/_appium/requirements.txt
Appium-Python-Client==5.1.1
pytest==7.4.0
pytest-dependency==0.60
pytest-order==1.3.0
python-dotenv==1.0.1
allure-pytest==2.14.3
pytest-steps==1.8.0